+++ drivers/gpu/nvgpu/gk20a/mm_gk20a.h
@@ -220,7 +220,7 @@ struct gk20a_mmu_level {
 			   struct scatterlist **sgl,
 			   u64 *offset,
 			   u64 *iova,
-			   u32 kind_v, u32 *ctag,
+			   u32 kind_v, u64 *ctag,
 			   bool cacheable, bool unmapped_pte,
 			   int rw_flag, bool sparse, bool priv);
 	size_t entry_size;
